removed unnecessary code and added some code

+++ app/Http/Controllers/DepartmentController.php
@@ -12,32 +12,37 @@ class DepartmentController extends Controller
         $this->middleware('auth');
     }
     public function index()
-    { 
-        $department = Department::all() ;
+    {
+        if (auth()->user()->role_id != 1) {
+            return redirect()->back()->with('error', 'You are not authorize to access the page');
+        }
+        $department = Department::all();
         return view('department.index', compact('department'));
     }
 
     public function destroy(Department $department)
     {
+        if ($department->users()->count() > 0) {
+            return redirect(route('departments.index'))->with('error', 'Cannot Delete Department because of assiociated employee');
+        }
         $department->delete();
         return redirect(route('departments.index'))->with('message', 'department deleted');
     }
 
-
-
     public function store(Request $request)
-    { 
-            Department::create([ 'name'=> $request->name ]);
-            return redirect(route('departments.index'))->with('message', 'New Department added'); 
-    } 
- 
-    public function show()
-    {  
+    {
+        $request->validate([
+            'name' =>  ['required'],
+        ]);
+        Department::create(['name' => $request->name]);
+        return redirect(route('departments.index'))->with('message', 'New Department added');
     }
- 
+
     public function update(Department $department, Request $request)
     {
- 
+        $request->validate([
+            'name' =>  ['required'],
+        ]);
         $department->update($request->all());
         return redirect()->back()->with('message', 'Edit Saved!');
     }
